feat(travel): indicate inconsistent connections, Trips page
Visits are ordered by weight of orderInTrip property, because other
options is unreliable (like date of visit) or unreliable and cumbersome
(like recurrent reference of arrival and departure rides). Both,
dates and rides references could be not specified by user, it's ok

But because of this, there is no guarantee, that dates or
rides references will be consistent with order of visits.

This feature aim is to indicate for user such inconsistencies.
Particularly for rides references inconsistency