feat(travel): add visit and edit visit buttons, make add button sortable
Add button is sortable, to be able to insert new visit in arbitrary
position between existent visits. Inject dummy item in visits list
for that and wrap visit component to intercept it. This is required
because react-sortable-hoc expect all sortable items to be created
through same component, wrapped by HOC