fix(request-kit): if identity isn't specified, run request once on mount
To make identity optional.

Considered alternative solutions:
1. Never request without identity flag.
Solid, but then identity will be required. It would be desirable to
avoid it, because otherwise this crafty concept will be mandatory for
lib usage, and without necessity increase learning curve.
1. Make equal check for whole requirements.
This is error prone approach with impact on performance. Equal check
could return false negative and strike a lot of unnecessary requests,
or false positives and keep app in obsolete state. Both is unacceptable

Possible worth to do an additional check, that use "request on mount"
strategy, only if identity isn't present in object, not just undefined.
Seems to be more tricky, but use same approach like isMissingIf flag.
It could prevent unintentional request trigger, and allow undefined to
be an identity value (no exceptional cases)