fix(travel): initial input value in VisitEditCard suggest
+ remove unused dicts

After last changes only locations for trip originLocationId is fetched,
not all. So there is no data in location dict, to resolve location name.
Instead location name could be obtained from visit model, and it value
was already present in initialState value, but was ignored in customHook

Fix to use this value