feat(request-kit): allow to filter keys in mergeProvisionState
selectDomainsState will return state/emptyState for each domain
in hierarchy. And if at least one empty state will be merged, then
isComplete flag will be false.

But for example when merge domain tripPage, to check is tripPage.trips
and tripPage.visits and etc. is complete, tripPage domain also will be
included with empty state. So this is escape hatch, to exclude it from
merge