feat(request-kit): change state structure, add selectIsValid selector
Add isError, isValid flags, as they also could be in combination.

Reduce result only to one field: lastResult & error only to lastError,
but indicate is it actual or not through flags:

if isError or !isValid then lastResult is not actual
if !isError then lastError is not actual

Actual means that data in result is reliable, otherwise it could only be
used as placeholder, during update. Or should be marked as obsolete,
for example by gray font color or background