refactor(app): navigation configuration in one place
Navigation is in app area of responsibility, not packages. It's more
convenient to keep all site map in one place

It's not final solution. Next step is to eliminate deprecated and dead
code related to navigation related code in packages.

But navigation != routing. Packages provides some routes, but app
configure navigation based on them. Because app knows all it's packages,
but packages doesn't knows app