refactor: definePackages
Only define plain config (with mountPath) instead of creation of package
Creation is complex and not needed in current app structure.

Also pass aliases within package definitions. In general same package
could be used several times with different mountPath (such cases exists)
But package name still be the same, thus need to introduce direct alias
for mountPath.

Alias is used for ease access to package details, because mountPath
could be empty '' or have slashes or even parameters and it inconvenient
to use mountPath as object key